requirements.txt
PyYAML
locust
names
requests
pytest
pytest-html
selenium
webdriver_manager